👌 Rename `formats.ts` to `format.ts` to match the styleguide
